Revert "Modulo Personas"
This reverts commit 73672586dc47838644c203722d595f640d09fe50.
