+++ src/Estructura/Persona.java
@@ -3,38 +3,62 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Personas;
+package Estructura;
 
 import java.util.Date;
 
 /**
  *
- * @author Hugo
+ * @author Edwin Chocoy
  */
 public interface Persona {
-
+    
+    /**
+     *
+     * @param nombre
+     */
     public void setNombre(String nombre);
     
+    public String getNombre();
+    
+    /**
+     *
+     * @param apellido
+     */
     public void setApellido(String apellido);
     
+    public String getApellido();
+    
+    /**
+     *
+     * @param dpi
+     */
     public void setDPI(String dpi);
     
+    public String getDPI();
+    
+    /**
+     *
+     * @param genero
+     */
     public void setGenero(boolean genero);
     
+    public boolean getGenero();
+    
+    /**
+     *
+     * @param fecha
+     */
     public void setFechaNacimiento(Date fecha);
     
+    public Date getFechaNacimiento();
+    
+    /**
+     *
+     * @param telefono
+     */
     public void setNumeroTelefonico(String telefono);
     
-    public String getNombre();
-
-    public String getApellido();
-
-    public String getDPI();
-
-    public boolean getGenero();
-
-    public Date getFechaNacimiento();
-
     public String getNumeroTelefonico();
     
     public int calcularEdad();
