+++ src/Estructura/Trabajador.java
@@ -5,7 +5,6 @@
  */
 package Estructura;
 
-import Personas.Persona;
 import java.util.ArrayList;
 
 /**
