--- src/Personas/AdaptadorEstructura.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Personas;
-
-/**
- *
- * @author Hugo
- */
-public interface AdaptadorEstructura {
-    public Persona getPersona(Integer idEmpleado);
-}
