--- src/Personas/exceptions/NonexistentEntityException.java
@@ -1,10 +1,0 @@
-package Personas.exceptions;
-
-public class NonexistentEntityException extends Exception {
-    public NonexistentEntityException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    public NonexistentEntityException(String message) {
-        super(message);
-    }
-}
